- moved dragToSelect to own component
